I changed the serialization format for the objects in MITIE-models.
I also added a get_fingerprint() method to the objects so that we can
version their contents.  This way other models that depend on each other
can identify which version of each model they are trained to use.

Finally, each object in MITIE-models now has a serialization format that
begins with a string identifying the name of the class in the file and
then the serialized object itself.